@@ -13,8 +13,12 @@
     "http-proxy": "bin/http-proxy"
   },
   "scripts": {
+    "cli": "ts-node src/cli.ts",
+    "cli:debug": "npm run compile && node --inspect dist/cli.js",
+    "format": "prettier --write 'src/**/*.ts'",
+    "compile": "rm -rf dist && tsc",
     "test": "echo \"Error: no test specified\" && exit 1",
-    "prepublish": "rm -rf dist && tsc"
+    "prepublish": "npm run compile"
   },
   "repository": {
     "type": "git",
@@ -32,23 +36,25 @@
   "homepage": "https://github.com/leizongmin/leizm-http-proxy#readme",
   "dependencies": {
     "@types/cli-color": "^0.3.29",
-    "@types/debug": "0.0.29",
-    "@types/fs-extra": "^3.0.3",
+    "@types/debug": "0.0.30",
+    "@types/fs-extra": "^5.0.2",
     "@types/js-yaml": "^3.5.31",
-    "@types/mime": "^1.3.1",
-    "@types/node": "^8.0.1",
+    "@types/mime": "^2.0.0",
+    "@types/node": "^10.0.2",
     "@types/path-to-regexp": "^1.7.0",
-    "@types/yargs": "^6.6.0",
+    "@types/yargs": "^11.0.0",
     "cli-color": "^1.2.0",
-    "debug": "^2.6.8",
-    "fs-extra": "^3.0.1",
+    "debug": "^3.1.0",
+    "fs-extra": "^6.0.0",
     "js-yaml": "^3.8.4",
-    "mime": "^1.3.6",
-    "path-to-regexp": "^1.7.0",
-    "source-map-support": "^0.4.15",
-    "yargs": "^8.0.2"
+    "mime": "^2.3.1",
+    "path-to-regexp": "^2.2.1",
+    "source-map-support": "^0.5.5",
+    "yargs": "^11.0.0"
   },
   "devDependencies": {
-    "typescript": "^2.3.4"
+    "prettier": "^1.12.1",
+    "ts-node": "^6.0.2",
+    "typescript": "^2.8.3"
   }
 }
