@@ -1,13 +1,13 @@
-import HTTPProxy from './';
+import HTTPProxy from "./";
 
 const proxy = new HTTPProxy();
 
 proxy.addRule({
-  match: 'http://morning.work/*',
-  proxy: 'http://ucdok.com/{1}',
+  match: "http://morning.work/*",
+  proxy: "http://ucdok.com/{1}",
   headers: {
-    host: 'jsxss.com',
+    host: "jsxss.com",
   },
 });
 
-proxy.server.listen(4567, () => console.log('listening...'));
+proxy.server.listen(4567, () => console.log("listening..."));
